rename filewriter.py to files.py
